Target specific version of .NET Core SDK and Runtime to prevent caching issues
Ran into a problem where my local machine used an older version of dotnet CLI with a specific behavior for publishing apps, and Azure DevOps pulled in newer versions with a different behavior. This broke the build on Azure DevOps but continued to work on my machine.

Specifically the issue was the older version of dotnet CLI would place the published app directory at the same level as the *.fsproj project file (so ./src/dist), but the newer version places it wherever the command is run from (so ./dist). The Dockerfile was looking at the older directory location, which worked on the cached version of fsharp:netcore image, but not on the uncached version (updated with newer dotnet CLI).